% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay_trip.R
\name{delay_trip}
\alias{delay_trip}
\title{Shift Trips in Time}
\usage{
delay_trip(gtfs, trip, duration)
}
\arguments{
\item{gtfs}{A GTFS object.}

\item{trip}{Character vector of `trip_id` values.}

\item{duration}{Numeric seconds or an object coercible to seconds, such as
a `difftime`.}
}
\value{
A modified `wizardgtfs` object.
}
\description{
Adds a number of seconds to every non-empty arrival and departure time for
selected trips. GTFS times above 24 hours are preserved.
}
\examples{
delayed <- delay_trip(
  for_rail_gtfs,
  trip = for_rail_gtfs$trips$trip_id[1:2],
  duration = 300
)

}
\seealso{
[GTFSwizard::edit_speed()], [GTFSwizard::set_dwelltime()]
}
