% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_speed.R
\name{edit_speed}
\alias{edit_speed}
\title{Scale In-Vehicle Travel Speed}
\usage{
edit_speed(gtfs, trips = "all", stops = "all", factor)
}
\arguments{
\item{gtfs}{A GTFS object.}

\item{trips}{Character trip IDs or `"all"`.}

\item{stops}{Character stop IDs or `"all"`. A segment is edited when either
endpoint is selected.}

\item{factor}{One positive speed multiplier. For example, `2` halves segment
travel times.}
}
\value{
A modified `wizardgtfs` object.
}
\description{
Changes travel time between consecutive stops by dividing it by a speed
multiplier. Dwell times are preserved and all downstream times are shifted.
}
\examples{
edited <- edit_speed(
  for_rail_gtfs,
  trips = for_rail_gtfs$trips$trip_id[1:2],
  stops = "all",
  factor = 1.25
)

}
\seealso{
[GTFSwizard::get_speeds()], [GTFSwizard::get_durations()]
}
