% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_1stdeparture.R
\name{get_1stdeparture}
\alias{get_1stdeparture}
\title{Get the First Departure of Each Trip}
\usage{
get_1stdeparture(gtfs)
}
\arguments{
\item{gtfs}{A GTFS object.}
}
\value{
A tibble with `route_id`, `trip_id`, `departure_time`, and
  `stop_id`.
}
\description{
Returns the departure at the lowest `stop_sequence` for every trip.
}
\examples{
head(get_1stdeparture(for_rail_gtfs))

}
\references{
[GTFS Schedule Reference](https://gtfs.org/documentation/schedule/reference/#stop_timestxt)
}
