% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hubs.R
\name{get_hubs}
\alias{get_hubs}
\title{Identify Stops Serving Multiple Routes}
\usage{
get_hubs(gtfs)
}
\arguments{
\item{gtfs}{A GTFS object.}
}
\value{
A WGS84 `sf` object with one row per served stop, list-columns
  `trip_id` and `route_id`, counts `n_trip` and `n_routes`, and point
  geometry.
}
\description{
Identify Stops Serving Multiple Routes
}
\examples{
get_hubs(for_rail_gtfs)

}
\seealso{
[GTFSwizard::plot_hubs()], [GTFSwizard::get_stops_sf()]
}
