% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shapes.R
\name{get_shapes}
\alias{get_shapes}
\title{Infer Straight-Line Shapes from Stop Sequences}
\usage{
get_shapes(gtfs)
}
\arguments{
\item{gtfs}{A GTFS object.}
}
\value{
A `wizardgtfs` object with a new `shapes` table and updated
  `trips$shape_id`.
}
\description{
Builds one shape for each unique ordered stop pattern and assigns it to the
corresponding trips.
}
\details{
Consecutive stop coordinates are connected directly. The result is useful
for visualization and approximate distance analysis, but it is not a
map-matched representation of the vehicle path.
}
\examples{
gtfs_with_shapes <- get_shapes(for_rail_gtfs)

}
\seealso{
[GTFSwizard::get_shapes_sf()], [GTFSwizard::get_shapes_df()]
}
