% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shapes_df.R
\name{get_shapes_df}
\alias{get_shapes_df}
\title{Convert Shape Geometries to a GTFS Shapes Table}
\usage{
get_shapes_df(shape)
}
\arguments{
\item{shape}{An `sf` object with a `shape_id` field and line geometries.}
}
\value{
A tibble containing `shape_id`, WGS84 point coordinates,
  `shape_pt_sequence`, and cumulative `shape_dist_traveled` in meters.
}
\description{
Convert Shape Geometries to a GTFS Shapes Table
}
\details{
Distances are geodesic approximations between consecutive coordinates.
GTFS permits any consistent distance unit; this function uses meters.
}
\examples{
shapes_sf <- get_shapes_sf(for_rail_gtfs)$shapes
shapes_df <- get_shapes_df(shapes_sf)

}
\references{
[GTFS Schedule Reference](https://gtfs.org/documentation/schedule/reference/#shapestxt)
}
\seealso{
[GTFSwizard::get_shapes_sf()], [GTFSwizard::get_shapes()]
}
