% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stops_sf.R
\name{get_stops_sf}
\alias{get_stops_sf}
\title{Convert GTFS Stops to Simple Features}
\usage{
get_stops_sf(gtfs)
}
\arguments{
\item{gtfs}{A GTFS object or a `stops` data frame.}
}
\value{
For a data frame, a WGS84 point `sf` object. For a GTFS object, the
  same object with its `stops` table converted to `sf`.
}
\description{
Convert GTFS Stops to Simple Features
}
\examples{
gtfs_sf <- get_stops_sf(for_rail_gtfs)
stops_sf <- get_stops_sf(for_rail_gtfs$stops)

}
