% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlon2epsg.R
\name{latlon2epsg}
\alias{latlon2epsg}
\title{Transform Spatial Data to a Local Metric CRS}
\usage{
latlon2epsg(sf_obj)
}
\arguments{
\item{sf_obj}{An `sf` or `sfc` object with a defined CRS.}
}
\value{
The input transformed to EPSG 326xx/327xx, EPSG 3413, or EPSG 3031.
}
\description{
Selects a UTM zone from the geographic centroid, or a polar stereographic
CRS outside UTM's latitude range.
}
\examples{
shapes <- get_shapes_sf(for_rail_gtfs)$shapes
metric_shapes <- latlon2epsg(shapes)

}
\seealso{
[sf::st_transform()]
}
