% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_.R
\name{plot_frequency}
\alias{plot_frequency}
\title{Plot System Frequency by Hour}
\usage{
plot_frequency(gtfs)
}
\arguments{
\item{gtfs}{A GTFS object.}
}
\value{
A `ggplot` object.
}
\description{
Shows the distribution and weighted mean of scheduled trip departures by
hour. Service-pattern date counts provide the weights.
}
\examples{
plot_frequency(for_rail_gtfs)

}
\seealso{
[GTFSwizard::get_frequency()]
}
