% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_.R
\name{plot_headways}
\alias{plot_headways}
\title{Plot System Headway by Hour}
\usage{
plot_headways(gtfs)
}
\arguments{
\item{gtfs}{A GTFS object.}
}
\value{
A `ggplot` object with headway in minutes.
}
\description{
Plot System Headway by Hour
}
\examples{
plot_headways(for_rail_gtfs)

}
\seealso{
[GTFSwizard::get_headways()]
}
