% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_service_planning.R
\name{plot_routeduration}
\alias{plot_routeduration}
\title{Plot Scheduled Trip Duration by Route}
\usage{
plot_routeduration(gtfs, top_n = 20L)
}
\arguments{
\item{gtfs}{A GTFS object.}

\item{top_n}{Maximum number of routes to display.}
}
\value{
A `ggplot` object. Each observation is one scheduled trip; boxes
  summarize duration in minutes by route.
}
\description{
Shows the distribution of scheduled end-to-end trip duration for the
busiest routes.
}
\examples{
plot_routeduration(for_rail_gtfs)

}
\references{
[FTA Evaluation Introduction](https://www.transit.dot.gov/research-innovation/evaluation-introduction)
}
