% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_.R
\name{plot_routefrequency}
\alias{plot_routefrequency}
\title{Plot Route Frequency by Hour}
\usage{
plot_routefrequency(gtfs, route = NULL)
}
\arguments{
\item{gtfs}{A GTFS object.}

\item{route}{Optional character vector of route IDs. `NULL` includes all
routes.}
}
\value{
A `ggplot` object.
}
\description{
Plot Route Frequency by Hour
}
\examples{
plot_routefrequency(
  for_rail_gtfs,
  route = for_rail_gtfs$routes$route_id[1:2]
)

}
\seealso{
[GTFSwizard::get_frequency()]
}
