% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_.R
\name{plot_routeheadways}
\alias{plot_routeheadways}
\title{Plot Route-Specific Headway by Hour}
\usage{
plot_routeheadways(gtfs, route = NULL)
}
\arguments{
\item{gtfs}{A GTFS object, ideally of the `wizardgtfs` class, or it will be converted.}

\item{route}{A character vector specifying the route ID(s) to plot. If `NULL`, all routes will be included.}
}
\value{
A `plotly` interactive plot showing hourly headway distributions for the selected routes, including:

  - Hourly Distribution: Boxplots of the headway distribution across hours.

  - Hourly Average Headway: A line showing the weighted average headway for each hour.

  - Overall Average Headway: A dashed line marking the system's overall average headway.
}
\description{
`plot_routeheadways` generates an interactive plot of the headways (time between trips) by hour for specified routes within the GTFS dataset. The plot includes the hourly headway distribution, hourly average headway, and an overall average headway line.
}
\details{
The function calculates hourly and overall average headways, weighting by `pattern_frequency`. It provides a visual analysis of headway patterns by hour, allowing insights into route-specific service frequency variations throughout the day.
}
\examples{
\dontrun{
# Plot the headway distribution by hour for specific routes
plot_routeheadways(for_gtfs, route = for_gtfs$trips$trip_id[1:2])
}

}
\seealso{
[GTFSwizard::get_headways()]
}
