% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_service_planning.R
\name{plot_serviceheatmap}
\alias{plot_serviceheatmap}
\title{Plot Average Scheduled Departures by Weekday and Hour}
\usage{
plot_serviceheatmap(gtfs)
}
\arguments{
\item{gtfs}{A GTFS object.}
}
\value{
A `ggplot` object. Each tile is a weekday-hour combination and its
  fill is the mean number of scheduled trip departures per active date.
}
\description{
Creates a heatmap of average scheduled departures on active service dates.
}
\examples{
plot_serviceheatmap(for_rail_gtfs)

}
\references{
[GTFS Schedule Reference](https://gtfs.org/documentation/schedule/reference/)
}
