% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_service_planning.R
\name{plot_servicesupply}
\alias{plot_servicesupply}
\title{Plot Scheduled Vehicle-Hours by Route}
\usage{
plot_servicesupply(gtfs, top_n = 20L)
}
\arguments{
\item{gtfs}{A GTFS object.}

\item{top_n}{Maximum number of routes to display.}
}
\value{
A `ggplot` object. Each bar is scheduled vehicle-hours for one
  route on one representative service-pattern day.
}
\description{
Sums scheduled end-to-end trip durations for each route and service
pattern.
}
\examples{
plot_servicesupply(for_rail_gtfs)

}
\references{
[FTA National Transit Database](https://www.transit.dot.gov/ntd)
}
