% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_dwelltime.R
\name{set_dwelltime}
\alias{set_dwelltime}
\title{Set Dwell Times}
\usage{
set_dwelltime(gtfs, duration = 30, trips = "all", stops = "all")
}
\arguments{
\item{gtfs}{A GTFS object.}

\item{duration}{One non-negative dwell time in seconds.}

\item{trips, stops}{Character ID vectors or `"all"`.}
}
\value{
A modified `wizardgtfs` object.
}
\description{
Sets dwell time at selected trip-stop calls and propagates each change to
all later times in the same trip. Arrival at the edited stop is retained.
}
\examples{
edited <- set_dwelltime(
  for_rail_gtfs,
  duration = 30,
  trips = for_rail_gtfs$trips$trip_id[1:2],
  stops = for_rail_gtfs$stops$stop_id[1:2]
)

}
\seealso{
[GTFSwizard::edit_dwelltime()], [GTFSwizard::get_dwelltimes()]
}
