% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_dwelltime.R
\name{edit_dwelltime}
\alias{edit_dwelltime}
\title{Edit Dwell Time in GTFS Data}
\usage{
edit_dwelltime(gtfs, value)
}
\arguments{
\item{gtfs}{A GTFS object, either as a list, a `wizardgtfs` object or a `wizardgtfs_selected` object, containing GTFS data.}

\item{value}{A named numeric vector of length 1 specifying how the dwell time should be modified.
 The values must be objects of the `lubridate::duration()` class or they will not be interpreted as seconds.
 The valid names are "set", "to", "add", and "mult". If `value` is not named, it defaults to "set".
\describe{
  \item{\code{`set`}}{Overwrites dwell time for all or selected trips. This method is suitable for adding a dwelltime to a trip where this parameter is zero, it maintains the total duration of the trip.}
  \item{\code{`to`}}{Sets dwell time to the specified value for selected trips.}
  \item{\code{`add`}}{Increases current dwell time by the specified amount at selected stops.}
  \item{\code{`mult`}}{Multiplies current dwell time by the specified factor at selected stops.}
}}
}
\value{
A modified wizardgtfs object with updated dwell times.
}
\description{
Modifies the dwell time of trips in a `wizardgtfs` object based on the specified value parameter.
}
\details{
The function modifies the `stop_times` table in the GTFS object by updating dwell times according to the provided `value` parameter.
If gtfs is a wizardgtfs_selected object the operation is only performed on the selection.
If `value` is specified as `"set"`, the dwell time is updated for the entire trip unless specific stops are selected.

A warning is displayed if `"set"` is used without proper selection to inform users that it affects entire trips.
Use a selection to modify dwell time only at specific stops or trips.

This method modifies the duration of trips, except when the value parameter is set to 'set'.
}
\examples{
gtfs <- for_bus_gtfs

modified_gtfs <- edit_dwelltime(gtfs, value = c(set = 10))

modified_gtfs <- gtfs \%>\%
  selection(route_id \%in\% gtfs$routes$route_id[1:2]) \%>\%
  edit_dwelltime(value = c(add = duration(15)))

}
\seealso{
[GTFSwizard::selection()], [GTFSwizard::unselection()],
}
