% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_calendar.R
\name{plot_calendar}
\alias{plot_calendar}
\title{Plot the GTFS Service Calendar}
\usage{
plot_calendar(gtfs, ncol = 4, facet_by_year = FALSE)
}
\arguments{
\item{gtfs}{A GTFS object.}

\item{ncol}{Number of facet columns when `facet_by_year = FALSE`.}

\item{facet_by_year}{Logical. Arrange years in rows and months in columns.}
}
\value{
A `ggplot` object.
}
\description{
Creates a calendar heatmap of scheduled trips by service date.
}
\examples{
plot_calendar(for_rail_gtfs, ncol = 4)

}
\seealso{
[GTFSwizard::get_servicepattern()]
}
