% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_.R
\name{plot_corridor}
\alias{plot_corridor}
\title{Plot High-Frequency Corridors}
\usage{
plot_corridor(gtfs, i = 0.01, min_length = 1500, ...)
}
\arguments{
\item{gtfs}{A GTFS object.}

\item{i}{Proportion of highest-frequency segments to retain.}

\item{min_length}{Minimum corridor length in meters.}

\item{...}{Supports the legacy argument `min.length`.}
}
\value{
A `ggplot` map.
}
\description{
Plot High-Frequency Corridors
}
\seealso{
[GTFSwizard::get_corridor()]
}
