% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_.R
\name{plot_routefrequency}
\alias{plot_routefrequency}
\title{Plot Route Frequency by Hour}
\usage{
plot_routefrequency(gtfs, route = NULL, top_n = 25L)
}
\arguments{
\item{gtfs}{A GTFS object.}

\item{route}{Optional character vector of route IDs. `NULL` includes all
routes after applying `top_n`.}

\item{top_n}{Maximum number of routes to display when `route` is `NULL`.}
}
\value{
A `ggplot` object.
}
\description{
Creates a heatmap of scheduled departures by route, hour, and service
pattern. The tile fill is the number of scheduled trips.
}
\examples{
plot_routefrequency(
  for_rail_gtfs,
  route = for_rail_gtfs$routes$route_id[1:2]
)

}
\seealso{
[GTFSwizard::get_frequency()]
}
