% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gtfs.R
\name{read_gtfs}
\alias{read_gtfs}
\title{Read and Convert GTFS Data to `wizardgtfs` Format}
\usage{
read_gtfs(file.path, files = NULL, quiet = TRUE, ...)
}
\arguments{
\item{file.path}{A string specifying the path to the GTFS zip file.}

\item{files}{A character vector indicating specific GTFS files to read. If `NULL`, all files in the GTFS feed will be read.}

\item{quiet}{A logical value indicating whether to suppress messages and warnings during the import. Defaults to `TRUE`.}

\item{...}{Additional arguments.}
}
\value{
An object of class `wizardgtfs`, containing GTFS data as a list of data frames.
}
\description{
`read_gtfs` reads a GTFS feed from a zip file and converts it into a `wizardgtfs` object, suitable for transit data analysis.
}
\details{
If no specific files are indicated, all GTFS files within the zip archive are read. After importing, the function converts the GTFS data into a `wizardgtfs` object, which is tailored for efficient handling and analysis of transit data.
}
\examples{
\dontrun{
# Read a GTFS file and convert it to a wizardgtfs object
gtfs_data <- read_gtfs("path/to/gtfs.zip")

# Read only specific files from the GTFS feed
gtfs_data <- read_gtfs("path/to/gtfs.zip", files = c("stops.txt", "routes.txt"))
}

}
\seealso{
[GTFSwizard::as_wizardgtfs()], [GTFSwizard::write_gtfs()]
}
